LPS-79312 Do the same with Message Boards
